vibeedit-backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Firebase
firebase-admin==6.3.0

# AI/ML
langchain==0.1.4
langchain-openai==0.0.5
langchain-google-genai==0.0.6
openai==1.10.0
google-generativeai==0.3.2
openai-whisper==20231117

# Video Processing
moviepy==1.0.3
ffmpeg-python==0.2.0
Pillow==10.2.0

# Cloud Storage
boto3==1.34.25
google-cloud-storage==2.14.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# HTTP
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

